feat(line): 지하철 노선을 생성하는 서비스 메서드를 구현했다.

+++ src/main/java/subway/line/CreateLineRequest.java
@@ -7,12 +7,13 @@ import lombok.Getter;
 public class CreateLineRequest {
   private final String name;
   private final String color;
-  private final Integer upStationId;
-  private final Integer downStationId;
+  private final Long upStationId;
+  private final Long downStationId;
   private final Integer distance;
 
   @Builder
-  public CreateLineRequest(String name, String color, Integer upStationId, Integer downStationId, Integer distance) {
+  public CreateLineRequest(
+      String name, String color, Long upStationId, Long downStationId, Integer distance) {
     this.name = name;
     this.color = color;
     this.upStationId = upStationId;
